Major Clean Up 2
A bunch of errors with the packages started popping up such as with tailwind, prettier, eslint, and formkit. Hence restarting the project and configs was necessary.

if anything is wrong look at the branch created before this. If no errors are found in 2 weeks the branch will be removed.

NOTE: Update packages less often or dont at all from now on